refactor(frontend): move renderToolResult to helpers directory
