+++ apps/frontend/src/pages/chat/components/MessageList/components/ToolExecutionCard/components/ResultSection/helpers/renderToolResult.tsx
